More cleaning up

+++ plash/chrootimages.py
@@ -1,3 +1,4 @@
+import errno
 import os
 import shutil
 import subprocess
@@ -9,48 +10,53 @@ TMP_DIR = '/tmp/data/tmp'
 MNT_DIR = '/tmp/data/mnt'
 
 
-def staple_layer(layers, layer_cmd):
+def staple_layer(layers, layer_cmd, rebuild=False):
     last_layer = layers[-1]
     layer_name = hashstr(' '.join(layers + [layer_cmd]).encode())
+    final_child_dst = os.path.join(last_layer, 'children', layer_name)
 
-    # os.mkdir(os.path.join(last_layer, layer_name)) # what if already exists?
-    try:
-        os.mkdir(os.path.join(last_layer, 'children', layer_name))
-    except FileExistsError:
-        pass
-    try:
-        os.mkdir(os.path.join(last_layer, 'children', layer_name, 'payload'))
-    except FileExistsError:
-        pass
-    try:
-        os.mkdir(os.path.join(last_layer, 'children', layer_name, 'children'))
-    except FileExistsError:
-        pass
-
-    new_layer = os.path.join(last_layer, 'children', layer_name, 'payload')
-
-    try:
+    if not os.path.exists(final_child_dst) or rebuild:
+        new_child = mkdtemp(dir=TMP_DIR)
+        new_layer = os.path.join(new_child, 'payload')
         os.mkdir(new_layer)
-    except FileExistsError:
-        pass
-    build_at = mkdtemp(dir=TMP_DIR)
-    mountpoint = mount(layers=(os.path.join(i, 'payload') for i in layers), write_dir=build_at)
+        os.mkdir(os.path.join(new_child, 'children'))
 
-    p = subprocess.Popen(['chroot', mountpoint, 'bash', '-ce', layer_cmd])
-    pid = p.wait()
-    umount(mountpoint)
-    assert pid == 0, 'Building returned non zero exit status'
-    if pid != 0:
-        print('non zero exit status code when building')
-        shutil.rmtree(build_at)
-        assert False
+        mountpoint = mount(layers=(os.path.join(i, 'payload') for i in layers), write_dir=new_layer)
+
+        p = subprocess.Popen(['chroot', mountpoint, 'bash', '-ce', layer_cmd], stdout=2, stderr=2)
+        pid = p.wait()
+        umount(mountpoint)
+        assert pid == 0, 'Building returned non zero exit status'
+        if pid != 0:
+            print('non zero exit status code when building')
+            shutil.rmtree(build_at)
+            assert False
+        else:
+            try:
+
+                # if rebuilding get rid of current build if there is one
+                # this could be done with one call with renameat2
+                if rebuild:
+                    try:
+                        os.rename(final_child_dst, os.path.join(mkdtemp(dir=TMP_DIR)))
+                    except FileNotFoundError as exc:
+                        pass
+                    except OSError as exc:
+                        if exc.errno != errno.ENOTEMPTY:
+                            raise
+
+
+                os.rename(new_child, final_child_dst)
+            except OSError as exc:
+                if exc.errno == errno.ENOTEMPTY:
+                    print('*** plash: this layer already exists builded and will not be replaced (layer: {})'.format(layer_name))
+                else:
+                    raise
+
     else:
-        try:
-            os.rename(build_at, new_layer)
-        except OSError as exc:
-            raise
-            # check if it already exists, if yes cleanup (rmtree) and continue
-    return layers + [os.path.abspath(os.path.join(new_layer, '..'))]
+        print('*** plash: cached layer'.format(layer_name))
+
+    return layers + [final_child_dst]
 
 
 def umount(mountpoint):
@@ -60,8 +66,7 @@ def umount(mountpoint):
 
 def mount(layers, write_dir):
     mountpoint = mkdtemp(dir=MNT_DIR)
-
-    p = subprocess.Popen([
+    cmd = [
         'mount',
         '-t',
         'aufs',
@@ -70,7 +75,9 @@ def mount(layers, write_dir):
             write_dir=write_dir,
             mountpoint=':'.join(i + '=ro' for i in layers)),
         'none',
-        mountpoint])
+        mountpoint]
+    # print(cmd)
+    p = subprocess.Popen(cmd)
     exit = p.wait()
     assert exit == 0, 'mounting got non zero exit status'
     return mountpoint
@@ -80,14 +87,15 @@ def fetch_base_image(image_name):
 
 
 # exported
-def build(image, layers, *, quiet_flag=False, verbose_flag=False, rebuild_flag=False, extra_mounts=[]):
+def build(image, layers, *, quiet_flag=False, verbose_flag=False, rebuild_flag=False):
     base = [image]
     for layer in layers:
-        base = staple_layer(base, layer)
+        base = staple_layer(base, layer, rebuild=rebuild_flag)
     return base
 
 
-def run(layers, cmd):
+def run(base, layer_commands, cmd, *, quiet_flag=False, verbose_flag=False, rebuild_flag=False, extra_mounts=[]):
+    layers = build(base, layer_commands, rebuild_flag=True)
     mountpoint = mount([os.path.join(i, 'payload') for i in layers], mkdtemp(dir=TMP_DIR))
     os.chroot(mountpoint)
     os.chdir('/')
@@ -98,6 +106,6 @@ if __name__ == '__main__':
     # print(staple_layer(['/tmp/data/layers/ubuntu'], 'touch a'))
     # print(staple_layer(staple_layer(['/tmp/data/layers/ubuntu'], 'touch a'), 'touch b'))
     # print(build('/tmp/data/layers/ubuntu', ['touch a', 'touch b', 'rm a']))
-    run(build('/tmp/data/layers/ubuntu', ['touch a', 'touch b', 'rm /a']), ['/bin/bash'])
+    run('/tmp/data/layers/ubuntu', ['touch a', 'touch b', 'rm /a'], ['/bin/bash'], rebuild_flag=True,)
 
 # [asdlfjladsf, asdfjlkasdf, adsfjkd, adsfj4]
